[Feat]: 타로 결과 페이지 추가 및 라우팅 설정
