+++ src/app.tsx
@@ -4,6 +4,7 @@ import { BrowserRouter, Route, Routes } from 'react-router';
 import TairoPage from './pages/TairoPage';
 import TairoTheme from './pages/TairoTheme';
 import PageLayout from './layout/PageLayout';
+import TairoResult from './components/Organism/TairoResult';
 
 function App() {
   useEffect(() => {
@@ -23,6 +24,7 @@ function App() {
         <Route element={<PageLayout />}>
           <Route path="/tairo" element={<TairoPage />} />
           <Route path="/tairo-theme" element={<TairoTheme />} />
+          <Route path="/tairo-result" element={<TairoResult />} />
         </Route>
       </Routes>
     </BrowserRouter>
